WIP refactor moderator actions
